Did more reformatting.
Went through the PEP 8 recommendations and changed for example the class name to camel case instead of snake case.

Lines 361-362: Stopped the total count for a subject being greater than the total count inputted by the user. If there are no more tasks it exits the loop, instead of repeating.